+++ nbproject/project.properties
@@ -58,7 +58,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=ShoppingList.Test
+main.class=shopping_list.Test
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
