+++ src/database_helpers/SQLOpener.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package DatabaseHelpers;
+package database_helpers;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
